Completed migration from the assignment repo.
I originally wrote this for an assignment and then decided that
this was good enough to publish in a standalone repo.

I couldn't really find quality example code online when I was
trying this out myself. So I decided to put this out there.
(note: this is just my personal opinion)

After migrating the code, I also split it up and improved various
aspects of the UX and documentation.

The original code I wrote can be found at:
https://github.com/Hypro999/network-programming-assignment-two

Signed-off-by: Hemanth V. Alluri <hdrive1999@gmail.com>